Fix issue where certificate subject and issuer are truncated
* The X509_NAME_oneline function was being used to extract the certificate issuer and subject. The function
is being provided a static buffer which is only 254 chars in size, meaning that if either string is larger they will
be truncated and essential information may be lost.
* Increasing the buffer size or using a dynamic buffer could solve the issue, however the X509_NAME_oneline
function is deprecated and its use is strongly discouraged. Instead, the issue can be fixed using the newer
X509_NAME_print_ex function which will output the full string in a standard format.
* Unicode data is escaped in the output string by default, so ASN1_STRFLGS_ESC_MSB is explicitly cleared to
force the output of UTF8 encoded data. This allows matching unicode byte sequences in a more natural
and expected manner.
* This change has the side effect of addressing VirusTotal/yara#1392